truck: move password service to common folder
