+++ src/modules/user/services/create-user.service.ts
@@ -6,7 +6,7 @@ import { IUserRepository } from '../interfaces/repository.interface';
 import { UserStatus } from '../interfaces/user-status.enum';
 import { CreateUserDto } from '../dto/create-user.dto';
 import { ICreateUser, IUser } from '../interfaces/user.interface';
-import { PasswordService } from './password.service';
+import { PasswordService } from '../../../common/services/password.service';
 import { EmailService } from '../../../common/services/email.service';
 import { User } from '@prisma/client';
 import { RedisCacheService } from '../../../common/redis/redis-cache.service';
