+++ src/modules/user/test/user.service.spec.ts
@@ -16,7 +16,7 @@ import {
   MockUser,
   MockUserData,
 } from './mocks/user.mock';
-import { PasswordService } from '../services/password.service';
+import { PasswordService } from '../../../common/services/password.service';
 import { EmailService } from '../../../common/services/email.service';
 import { RedisCacheService } from '../../../common/redis/redis-cache.service';
 import { Channel } from '@prisma/client';
